moving scheduler test to another unit test, see if travis works now
