+++ kuyruk/test/integration/test_kuyruk.py
@@ -232,51 +232,3 @@ class KuyrukTestCase(unittest.TestCase):
             worker.expect('Exiting')
             worker.expect_exit(0)
         assert is_empty('kuyruk'), worker.get_output()
-
-    def _test_scheduler(self):
-        """Scheduler schedules correctly"""
-        delete_queue('scheduled')
-
-        def get_message_count():
-            from kuyruk import Worker
-            k = Kuyruk()
-            w = Worker(kuyruk=k, queue_name='scheduled')
-            w.started = time()
-            return w.get_stats()['queue']['messages_ready']
-
-        if os.path.exists('/tmp/kuyruk-scheduler.db'):
-            os.unlink('/tmp/kuyruk-scheduler.db')
-
-        config = {
-            'SCHEDULE': {
-                'runs-every-5-seconds': {
-                    'task': 'kuyruk.test.tasks.scheduled',
-                    'schedule': timedelta(seconds=5),
-                    'args': ['hello world from scheduler']
-                }
-            },
-            'SCHEDULER_FILE_NAME': '/tmp/kuyruk-scheduler'
-        }
-
-        p = Process(target=run_scheduler, kwargs={'config': config})
-        p.start()
-        sleep(6)
-        p.terminate()
-        self.assertEqual(get_message_count(), 2)
-
-        # checking shelve, this shouldnt send a job
-        p = Process(target=run_scheduler, kwargs={'config': config})
-        p.start()
-        p.terminate()
-        self.assertEqual(get_message_count(), 2)
-
-        sleep(2)
-
-        # # restart again, now it should send a job
-        p = Process(target=run_scheduler, kwargs={'config': config})
-        p.start()
-        sleep(3)
-        p.terminate()
-        self.assertEqual(get_message_count(), 3)
-
-        sleep(2)
